Revert "Changed from int to uint16_t some for cycles"
This reverts commit 534370c45f3cf16698c7182ed8c68e7a6fad93d1.

+++ common.h
@@ -4,24 +4,24 @@
 #include <cstdint>
 
 extern const int pc1[56];
-extern const uint16_t pc1_bin[56];
+extern const int pc1_bin[56];
 
 extern const int pc2[48];
-extern const uint16_t pc2_bin[48];
+extern const int pc2_bin[48];
 
 extern const int initial_permutation[64];
-extern const uint16_t initial_permutation_bin[64];
+extern const int initial_permutation_bin[64];
 
 extern const int expansion_table[48];
-extern const uint16_t expansion_table_bin[48];
+extern const int expansion_table_bin[48];
 
 extern const int substitution_boxes[8][4][16];
 extern const uint64_t substitution_boxes_bin[8][4][16];
 
 extern const int permutation_tab[32];
-extern const uint16_t permutation_tab_bin[32];
+extern const int permutation_tab_bin[32];
 
 extern const int inverse_permutation[64];
-extern const uint16_t inverse_permutation_bin[64];
+extern const int inverse_permutation_bin[64];
 
 #endif //DES_DECRYPTION_COMMON_H
